[chore] gfx layer adjustments (#3323)
* [chore] move common routines to gfx base

* [chore] remove device pointer inside every gfx resources

* [chore] file structure

* gfx: empty backend

* [chore] device creator abstraction

+++ native/cocos/bindings/event/EventDispatcher.cpp
@@ -94,7 +94,7 @@ void EventDispatcher::dispatchTouchEvent(const struct TouchEvent &touchEvent) {
         _jsTouchObjArray->root();
     }
 
-    _jsTouchObjArray->setProperty("length", se::Value(touchEvent.touches.size()));
+    _jsTouchObjArray->setProperty("length", se::Value((uint32_t)touchEvent.touches.size()));
 
     while (_jsTouchObjPool.size() < touchEvent.touches.size()) {
         se::Object *touchObj = se::Object::createPlainObject();
